+++ stdlib/Dates/src/io.jl
@@ -529,6 +529,14 @@ end
 
 # show
 
+function Base.show(io::IO, p::P) where P <: Period
+    if get(io, :compact, false)
+        print(io, p)
+    else
+        print(io, P, '(', p.value, ')')
+    end
+end
+
 function Base.show(io::IO, dt::DateTime)
     if get(io, :compact, false)
         print(io, dt)
@@ -539,9 +547,9 @@ function Base.show(io::IO, dt::DateTime)
         s = second(dt)
         ms = millisecond(dt)
         if ms == 0
-            print(io, "DateTime($y, $m, $d, $h, $mi, $s)")
+            print(io, DateTime, "($y, $m, $d, $h, $mi, $s)")
         else
-            print(io, "DateTime($y, $m, $d, $h, $mi, $s, $ms)")
+            print(io, DateTime, "($y, $m, $d, $h, $mi, $s, $ms)")
         end
     end
 end
@@ -559,7 +567,7 @@ function Base.show(io::IO, dt::Date)
         print(io, dt)
     else
         y,m,d = yearmonthday(dt)
-        print(io, "Date($y, $m, $d)")
+        print(io, Date, "($y, $m, $d)")
     end
 end